+++ Matilda/slovak_checkers/board/detail/common.h
@@ -8,13 +8,13 @@ namespace sc
     namespace detail
     {
         // one side of the checkers board
-        const constexpr size_t BOARD_SIZE = 8;
+        const constexpr uint8_t BOARD_SIZE = 8;
 
         // number of black squares, squares where pieces can be
-        const constexpr size_t SQUARES_COUNT = BOARD_SIZE * BOARD_SIZE / 2;
+        const constexpr uint8_t SQUARES_COUNT = BOARD_SIZE * BOARD_SIZE / 2;
 
         // invalid position on board
-        const constexpr size_t INVALID_POS = std::numeric_limits<size_t>::max();
+        const constexpr uint8_t INVALID_POS = std::numeric_limits<uint8_t>::max();
 
         using BitBoard = std::bitset<SQUARES_COUNT>;
     }
